requirements.txt
python-telegram-bot==20.0
Flask==2.3.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
gunicorn==20.1.0
flask-sqlalchemy==3.0.3